requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
Django==4.2.2
django-auto-logout==0.5.1
django-ckeditor==6.5.1
django-compat==1.0.15
django-crispy-forms==2.0
django-image-optimizer==1.0.3
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.4
mysqlclient==2.1.1
Pillow==9.5.0
py-dotenv==0.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-resize-image==1.1.20
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinify==1.6.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
